Create operate and add listener to equal button
Operate button that looks for whatever the operation string was and
whatever it is, do the said function and return the result. Then add
an event listener to the equals button that once it is pressed, we have
our second number which we convert that and the first number to numbers
and then call the operate function and display the result on the
display. Add classes to equals and clear button so can select them
later.