Add variable operatorPressedOnce
First set operatorPressedOnce to false, if an operator button is
clicked, then set it true, and there's an if statement saying that
if it was already true, then we have two numbers that we can evaluate.
We also move the clearing of the display by setting clearOnNextNum to
true when an operator button is pressed, so when a new number is
pressed, then the if statement clearOnNextNum is run and so there's
a blank board for the second number to be read.