Reorganize buttons and fill remaining space
Flex-grow set to 1 for all buttons so they fill equally the remaining
space they have after having a basis of 1/3 of the area. Then row
reverse the grid so numbers are forming from bottom to top like a
normal keyboard. Then adjust other buttons to account for the
row reverse