Round numbers if necessary and add overflow check
Add overflow checks that numbers will not be typed into display if it
exceed past 16 digits, and if the solution is over 16 digits, then the
display will shift to making it two significant figures. Also round
numbers to two decimal places only if necessary using +.toFixed.
Overflow checks is from if statements that check length